Travis CI: Test against fred branch master first, in preparation of running deployment only there
Currently deployment is supposed to happen on openjdk7, and now that we
test both against next and master on openjdk7, it would happen twice.
We only want to to happen once because uploads to Freenet consume a lot of
time and we don't want to abuse the free service of Travis.
So the following commits will change deployment to only happen on branch
master.
But to ensure that build job does not delay the whole Travis build it
should be started first, which this commit achieves.