Seed identities: Add new seeds to speed up announcement to active IDs
In Tests with the old Seed-IDs it took 5 hours for an ID to find the
first CAPTCHA / puzzle to solve, and that belonged to an ID that is
not active on Sone, giving no direct visibility in Sone.

In contrast the new Seed-IDs yield a CAPTCHA within roughly one hour,
so people who are active for two hours still have a chance to not only
be seen but to get a reply on their first message in their first
session.

The seednodes are a broad selection of longterm active IDs.