GitHub Actions: Temporarily copy GHA script from branch next to test branch build0021-development
Copied the file as of its commit
  1351cdd732adae82d2131e56ca03eb1dad4b9e08
on branch next.

NOTICE: This does not preserve the Git history!
Thus it should be reverted before merging branch next!

Doing this because branch build0021-development does not have a GitHub
Actions script yet, it is still on Travis CI, and merging the
sub-branches of next which would convert to GHA would be too much work
for now.